+++ src/errors.ts
@@ -1,21 +1,11 @@
-import {Domain, generators} from "alpha-errors";
+import {Domain, ErrorDescriptor, formatCodeFactory} from "@pallad/errors";
 
-export const ERRORS = Domain.create({
-    codeGenerator: generators.formatCode('E_RANGE_%d')
-})
-    .createErrors(create => {
-        return {
-            START_GREATER_THAN_END: create({
-                message: '"start" cannot be greater than "end"',
-                errorClass: TypeError
-            }),
-            UNDEFINED_BOUNDARIES: create({
-                message: 'Cannot create Range from undefined or null values',
-                errorClass: TypeError,
-            }),
-            EMPTY_ARRAY_ARGUMENT: create({
-                message: 'Cannot create range from empty array',
-                errorClass: TypeError
-            })
-        }
-    })
+const code = formatCodeFactory("E_RANGE_%c");
+
+export const errorsDomain = new Domain();
+
+export const ERRORS = errorsDomain.addErrorsDescriptorsMap({
+    START_GREATER_THAN_END: ErrorDescriptor.useDefaultMessage(code(1), '"start" cannot be greater than "end"', TypeError),
+    UNDEFINED_BOUNDARIES: ErrorDescriptor.useDefaultMessage(code(2), 'Cannot create Range from undefined or null values', TypeError),
+    EMPTY_ARRAY_ARGUMENT: ErrorDescriptor.useDefaultMessage(code(3), 'Cannot create range from empty array', TypeError)
+});